requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Models
pydantic==2.5.0

# RAG - Local Embeddings
sentence-transformers==2.3.1
huggingface-hub==0.20.3
scikit-learn==1.3.2
numpy==1.24.3
torch==2.1.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Vector Database
chromadb==0.4.22

httpx==0.27.0

sentence-transformers==2.3.1
